requirements.txt
requests==2.31.0
python-dotenv==1.0.0
urllib3==2.0.7
aiohttp==3.9.3
asyncio==3.4.3
